# apis/app-api/catalog/catalog-app-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Catalog App API",
    "version": "1.0.0",
    "description": "Typed app API for active catalog browsing, cart management, and user delivery addresses.",
    "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
    "x-sdkwork-sdk-family": "sdkwork-catalog-app-sdk",
    "x-sdkwork-owner": "sdkwork-catalog"
  },
  "servers": [{ "url": "/" }],
...
  "paths": {
    "/app/v3/api/catalog/attributes": {
      "get": {
        "tags": ["catalog"],
        "summary": "List active catalog attributes.",
        "operationId": "attributes.list",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/CatalogAttributeListSuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "catalog.attributes",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-pagination-mode": "offset"
      }
    },
    "/app/v3/api/catalog/categories": {
      "get": {
        "tags": ["catalog"],
        "summary": "List active catalog categories.",
        "operationId": "categories.list",
        "parameters": [
          { "name": "parent_id", "in": "query", "required": false, "schema": { "type": "string" } },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/CatalogCategoryListSuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "catalog.categories",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-pagination-mode": "offset"
      }
    },
    "/app/v3/api/catalog/categories/{categoryId}": {
      "get": {
        "tags": ["catalog"],
        "summary": "Retrieve an active catalog category.",
        "operationId": "categories.retrieve",
        "parameters": [{ "$ref": "#/components/parameters/CategoryId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/CatalogCategorySuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "catalog.categories",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api"
      }
    },
    "/app/v3/api/catalog/products": {
      "get": {
        "tags": ["catalog"],
        "summary": "List active catalog products.",
        "operationId": "products.list",
        "parameters": [
          { "name": "category_id", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "product_type", "in": "query", "required": false, "schema": { "type": "string" } },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/CatalogProductListSuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "catalog.products",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-pagination-mode": "offset"
      }
    },
    "/app/v3/api/catalog/products/{productId}": {
      "get": {
        "tags": ["catalog"],
        "summary": "Retrieve an active catalog product.",
        "operationId": "products.retrieve",
        "parameters": [{ "$ref": "#/components/parameters/ProductId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/CatalogProductSuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "catalog.products",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api"
      }
    },
    "/app/v3/api/catalog/products/{productId}/skus": {
      "get": {
        "tags": ["catalog"],
        "summary": "List active SKUs for a catalog product.",
        "operationId": "products.skus.list",
        "parameters": [
          { "$ref": "#/components/parameters/ProductId" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/CatalogSkuListSuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "catalog.products.skus",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-pagination-mode": "offset"
      }
    },
    "/app/v3/api/catalog/skus/{skuId}": {
      "get": {
        "tags": ["catalog"],
        "summary": "Retrieve an active catalog SKU.",
        "operationId": "skus.retrieve",
        "parameters": [{ "$ref": "#/components/parameters/SkuId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/CatalogSkuSuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "catalog.skus",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api"
      }
    },
    "/app/v3/api/cart/items": {
      "get": {
        "tags": ["cart"],
        "summary": "List the authenticated user's cart items.",
        "operationId": "items.list",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/CartItemListSuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "cart.items",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-pagination-mode": "offset"
      },
      "post": {
        "tags": ["cart"],
        "summary": "Add an item to the authenticated user's cart.",
        "operationId": "items.create",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddCartItemRequest" } } }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/CartItemSuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "cart.items",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api"
      }
    },
    "/app/v3/api/cart/items/{cartItemId}": {
      "patch": {
        "tags": ["cart"],
        "summary": "Update a cart item quantity.",
        "operationId": "items.update",
        "parameters": [{ "$ref": "#/components/parameters/CartItemId" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateCartItemRequest" } } }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/CartItemSuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "cart.items",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api"
      },
      "delete": {
        "tags": ["cart"],
        "summary": "Delete a cart item.",
        "operationId": "items.delete",
        "parameters": [{ "$ref": "#/components/parameters/CartItemId" }],
        "responses": {
          "204": { "description": "Cart item deleted." },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "cart.items",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api"
      }
    },
    "/app/v3/api/addresses": {
      "get": {
        "tags": ["delivery"],
        "summary": "List the authenticated user's delivery addresses.",
        "operationId": "addresses.list",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/AddressListSuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "addresses",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-pagination-mode": "offset"
      },
      "post": {
        "tags": ["delivery"],
        "summary": "Create a delivery address.",
        "operationId": "addresses.create",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateAddressRequest" } } }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/AddressSuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "addresses",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api"
      }
    },
...
    "/app/v3/api/addresses/{addressId}/default_selection": {
      "put": {
        "tags": ["delivery"],
        "summary": "Set the default delivery address.",
        "operationId": "addresses.defaultSelection.update",
        "parameters": [{ "$ref": "#/components/parameters/AddressId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/AddressSuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "addresses.defaultSelection",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api"
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AuthToken": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "AccessToken": { "type": "apiKey", "in": "header", "name": "Access-Token" }
    },
    "parameters": {
      "Page": { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "default": 1 } },
      "PageSize": { "name": "page_size", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 20 } },
      "CategoryId": { "name": "categoryId", "in": "path", "required": true, "schema": { "type": "string" } },
      "ProductId": { "name": "productId", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "responses": {
      "CatalogAttributeListSuccess": { "description": "Catalog attribute page.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogAttributeListResponse" } } } },
      "CatalogCategoryListSuccess": { "description": "Catalog category page.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogCategoryListResponse" } } } },
      "CatalogCategorySuccess": { "description": "Catalog category.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogCategoryResponse" } } } },
      "CatalogProductListSuccess": { "description": "Catalog product page.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogProductListResponse" } } } },
      "CatalogProductSuccess": { "description": "Catalog product.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogProductResponse" } } } },
      "CatalogSkuListSuccess": { "description": "Catalog SKU page.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogSkuListResponse" } } } },
...
      "BadRequest": { "description": "Invalid request.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } } },
      "Unauthorized": { "description": "Authentication required.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } } },
      "NotFound": { "description": "Resource not found.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } } },
      "InternalError": { "description": "Internal server error.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } } }
    },
...
  },
  "security": [{ "AuthToken": [], "AccessToken": [] }]
}